requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-socketio==5.1.1
numpy>=1.19.3
opencv-python-headless==4.5.3.56
pyyaml==5.4.1
python-engineio==4.2.1
python-socketio==5.4.0
setuptools>=57.0.0
wheel>=0.36.2
